+++ src/gen/empty_const.rs
@@ -380,27 +380,6 @@ mod tests {
     }
 
     #[test]
-    fn size_arr_vec_weighted() {
-        assert_eq!(<[Vec<(usize, usize)>; 1]>::empty().size(), 0);
-        assert_eq!(<[Vec<(usize, usize)>; 2]>::empty().size(), 0);
-        assert_eq!(<[Vec<(usize, usize)>; 3]>::empty().size(), 0);
-    }
-
-    #[test]
-    fn size_arr_btree_set_weighted() {
-        assert_eq!(<[BTreeSet<(usize, usize)>; 1]>::empty().size(), 0);
-        assert_eq!(<[BTreeSet<(usize, usize)>; 2]>::empty().size(), 0);
-        assert_eq!(<[BTreeSet<(usize, usize)>; 3]>::empty().size(), 0);
-    }
-
-    #[test]
-    fn size_arr_hash_set_weighted() {
-        assert_eq!(<[HashSet<(usize, usize)>; 1]>::empty().size(), 0);
-        assert_eq!(<[HashSet<(usize, usize)>; 2]>::empty().size(), 0);
-        assert_eq!(<[HashSet<(usize, usize)>; 3]>::empty().size(), 0);
-    }
-
-    #[test]
     fn size_arr_btree_map() {
         assert_eq!(<[BTreeMap<usize, usize>; 1]>::empty().size(), 0);
         assert_eq!(<[BTreeMap<usize, usize>; 2]>::empty().size(), 0);